--- src/index.js
@@ -1,3 +1,0 @@
-export * from './js-remote';
-export * from './ngJS-remote';
-export * from './ngJS-remote.service';
